+++ libs/causality/src/lib/CasualAnalysisDashboard/Controls/CasualAnalysisView/CasualAggregateView/CasualAggregateStyles.ts
@@ -2,38 +2,25 @@
 // Licensed under the MIT License.
 
 import {
-  FontWeights,
-  getTheme,
   IProcessedStyleSet,
   IStyle,
   mergeStyleSets
 } from "office-ui-fabric-react";
 
 export interface ICasualAggregateStyles {
-  callout: IStyle;
   container: IStyle;
   description: IStyle;
-  infoButton: IStyle;
   label: IStyle;
   lasso: IStyle;
-  link: IStyle;
   table: IStyle;
-  title: IStyle;
   leftPane: IStyle;
   rightPane: IStyle;
-  modalContentHelp: IStyle;
-  modalContentHelpText: IStyle;
 }
 
 export const CasualAggregateStyles: () => IProcessedStyleSet<
   ICasualAggregateStyles
 > = () => {
-  const theme = getTheme();
   return mergeStyleSets<ICasualAggregateStyles>({
-    callout: {
-      padding: "20px 24px",
-      width: 320
-    },
     container: {
       display: "flex",
       flex: 1,
@@ -43,19 +30,6 @@ export const CasualAggregateStyles: () => IProcessedStyleSet<
       display: "flex",
       justifyContent: "space-between",
       padding: "10px"
-    },
-    infoButton: {
-      borderRadius: "50%",
-      color: theme.semanticColors.bodyText,
-      float: "left",
-      fontSize: "12px",
-      fontWeight: "600",
-      height: "15px",
-      lineHeight: "14px",
-      marginRight: "3px",
-      marginTop: "3px",
-      textAlign: "center",
-      width: "15px"
     },
     label: {
       cursor: "pointer",
@@ -77,28 +51,11 @@ export const CasualAggregateStyles: () => IProcessedStyleSet<
       padding: "10px",
       width: "70%"
     },
-    link: {
-      display: "block",
-      marginTop: 20
-    },
-    modalContentHelp: {
-      float: "left",
-      width: "350px"
-    },
-    modalContentHelpText: {
-      padding: "0px 20px",
-      textAlign: "center",
-      wordWrap: "break-word"
-    },
     rightPane: {
       width: "25%"
     },
     table: {
       width: "50%"
-    },
-    title: {
-      fontWeight: FontWeights.semilight,
-      marginBottom: 12
     }
   });
 };
